\name{biogeme-package}
\alias{biogeme-package}
\alias{biogeme}
\docType{package}
\title{
\packageTitle{biogeme}
}
\description{
\packageDescription{biogeme}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{biogeme}
\packageIndices{biogeme}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{biogeme}

Maintainer: \packageMaintainer{biogeme}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
